Add the brep2graph python library
One of the ways how to apply machine learning tools for
a CAD data is to represent the CAD data (model) as a graph
and to apply ML tools to the graph (GNN).

This library collects different configurations how you
can reprensent the CAD (in step format) model as graph.